requirements.txt
pillow
python-multipart
fastapi[all]
ftfy
scipy
torch
torchvision
